Proper loading of alpha value Alpha value in colors is stored as 8 bit hex value. In the conversion to 16 bit there was a lack of proportionality, due to the fact that the 8 bit value was simply left shifted, eg. 0xab -> 0xab00. The new code duplicates the value (eg. 0xab -> 0xabab). Maybe a better approach would be using the 8 bit value directly, but the gtk color buttons are returning 16 bit alpha values.

+++ multiload/multiload.h
@@ -87,7 +87,7 @@ struct _LoadGraph {
 struct _GraphConfig {
 	gboolean visible;
 	GdkColor colors[MAX_COLORS];
-	guint alpha[MAX_COLORS];
+	guint16 alpha[MAX_COLORS];
 };
 
 struct _MultiloadPlugin
